examples/python/requirements.txt
absl-py==1.4.0
ale-py==0.7.4
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
filelock==3.11.0
fonttools==4.39.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.53.0
gym==0.21.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pip==21.1.1
protobuf==4.22.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
setuptools==56.0.0
six==1.16.0
stable-baselines3==1.8.0
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
torch==2.0.0
tqdm==4.65.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
wheel==0.40.0
zipp==3.15.0
